+++ components/support/nimbus-fml/src/backends/frontend_manifest.rs
@@ -27,6 +27,7 @@ impl From<FeatureManifest> for ManifestFrontEnd {
             channels,
             includes: Default::default(),
             imports: Default::default(),
+            no_lint: Default::default(),
             features,
             legacy_types: None,
             types: Types { enums, objects },